refactor(js): modules utils
