+++ source/js/modules/utils/NumbersAnimation.js
