+++ source/js/modules/utils/animate-letters.js
