+++ source/js/modules/utils/reload-svg-animation.js
